Further tweaks to 2024:22

+++ src/main/java/com/rev/aoc/problems/y2024/D22.java
@@ -3,19 +3,17 @@ package com.rev.aoc.problems.y2024;
 import com.rev.aoc.framework.io.load.LoaderUtils;
 import com.rev.aoc.framework.problem.AocCoordinate;
 import com.rev.aoc.framework.problem.AocProblem;
-import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.ArrayDeque;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.PriorityQueue;
 import java.util.Queue;
 import java.util.Set;
 
 public final class D22 extends AocProblem<Long, Long> {
-    private static final long PRUNE_MOD = 1 << 24;
+    private static final int PRUNE_MOD = 1 << 24;
     private static final int PART_ONE_ITERATIONS = 2000;
 
     @Override
@@ -51,12 +49,12 @@ public final class D22 extends AocProblem<Long, Long> {
     private long computePriceChangeScores(final int seed,
                                           final Map<List<Integer>, Integer> runsAndScores,
                                           long bestScoreSoFar) {
-        long price = seed;
+        int price = seed;
         Queue<Integer> priceChanges = new ArrayDeque<>();
         Set<List<Integer>> visited = new HashSet<>();
         for (int i = 0; i < PART_ONE_ITERATIONS; i++) {
-            long nextPrice = hashOnce(price);
-            priceChanges.add((int) ((nextPrice % 10) - (price % 10)));
+            int nextPrice = hashOnce(price);
+            priceChanges.add((nextPrice % 10) - (price % 10));
             if (priceChanges.size() > 4) {
                 priceChanges.remove();
             }
@@ -65,7 +63,7 @@ public final class D22 extends AocProblem<Long, Long> {
                 if (!visited.contains(priceChangesCpy)) {
                     visited.add(priceChangesCpy);
                     int score = runsAndScores.getOrDefault(priceChangesCpy, 0);
-                    int newScore = score + ((int) (nextPrice % 10));
+                    int newScore = score + (nextPrice % 10);
                     if (newScore > bestScoreSoFar) {
                         bestScoreSoFar = newScore;
                     }
@@ -77,14 +75,14 @@ public final class D22 extends AocProblem<Long, Long> {
         return bestScoreSoFar;
     }
 
-    private long calculateHash(long secretNumber, long iterations) {
+    private int calculateHash(int secretNumber, int iterations) {
         for (int i = 0; i < iterations; i++) {
             secretNumber = hashOnce(secretNumber);
         }
         return secretNumber;
     }
 
-    private static long hashOnce(long secretNumber) {
+    private static int hashOnce(int secretNumber) {
         secretNumber = (secretNumber ^ (secretNumber << 6)) & (PRUNE_MOD - 1);
         secretNumber = (secretNumber ^ (secretNumber >> 5)) & (PRUNE_MOD - 1);
         secretNumber = (secretNumber ^ (secretNumber << 11)) & (PRUNE_MOD - 1);
